Recommendations
Louise, "An Analysis of Kickstarter Campaigns" was converted from Excel to Pdf, as the original file was to large to move in GitHub. Revert back to Excel to make the content easier to understand. The first tab contains all information you have given. While, the following tabs have a detailed examination of the elements and structure of that information you have given.